fix ui

+++ include/screens.h
@@ -18,13 +18,13 @@ namespace screens
 }
 
 
-inline void toLayoutSpace(u8& x, u8& y)
+inline void toLayoutSpace(int& x, int& y)
 {
 	using namespace screens;
 	switch (settings::layout) {
 		case d0:	return;
-		case d90:	{u8 c=x; x=dimX-y; y=c; return;}
+		case d90:	{int c = x; x = dimX - y; y = c; return;}
 		case d180:	x=dimX-x; y=dimY-y; return;
-		case d270:	{u8 c=x; x=y; y=dimY-c;}
+		case d270:	{int c = x; x = y; y = dimY - c;}
 	}
 }
